python-sdk/requirements.txt
requests>=2.28.0
pillow>=9.0.0
typing-extensions>=4.0.0